+++ lib/elixir/lib/protocol.ex
@@ -155,7 +155,7 @@ defmodule Protocol do
   # Builtin types.
   @doc false
   def builtin do
-    [ Tuple, Atom, List, BitString, Number,
+    [ Tuple, Atom, List, BitString, Integer, Float,
       Function, PID, Port, Reference, Any ]
   end
 
@@ -206,7 +206,8 @@ defmodule Protocol do
   defp impl_for(current, Atom, arg),      do: impl_with_fallback(Atom, :is_atom, current, Any, arg)
   defp impl_for(current, List, arg),      do: impl_with_fallback(List, :is_list, current, Any, arg)
   defp impl_for(current, BitString, arg), do: impl_with_fallback(BitString, :is_bitstring, current, Any, arg)
-  defp impl_for(current, Number, arg),    do: impl_with_fallback(Number, :is_number, current, Any, arg)
+  defp impl_for(current, Integer, arg),   do: impl_with_fallback(Integer, :is_integer, current, Any, arg)
+  defp impl_for(current, Float, arg),     do: impl_with_fallback(Float, :is_float, current, Any, arg)
   defp impl_for(current, Function, arg),  do: impl_with_fallback(Function, :is_function, current, Any, arg)
   defp impl_for(current, PID, arg),       do: impl_with_fallback(PID, :is_pid, current, Any, arg)
   defp impl_for(current, Port, arg),      do: impl_with_fallback(Port, :is_port, current, Any, arg)